requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
email-validator
httpx
pytest
pytest-asyncio
pytest-cov
python-dotenv
passlib[bcrypt]
pyjwt
